+++ apps/mobile/index.js
@@ -1,6 +1,5 @@
 import './shim.js';
 import 'react-native-gesture-handler';
-import 'react-native-reanimated';
 
 import { AppRegistry } from 'react-native';
 import { App } from 'ui';
