+++ apps/mobile/package.json
@@ -56,7 +56,6 @@
     "react-native-qrcode-scanner": "^1.5.5",
     "react-native-qrcode-svg": "^6.1.2",
     "react-native-randombytes": "^3.0.0",
-    "react-native-reanimated": "^2.13.0",
     "react-native-safe-area-context": "^4.2.4",
     "react-native-screens": "^3.18.0",
     "react-native-sha256": "^1.4.7",
