+++ libs/ui/jest.setup.js
@@ -5,6 +5,5 @@ jest.mock('react-native/Libraries/Animated/NativeAnimatedHelper');
 jest.mock('react-native/Libraries/EventEmitter/NativeEventEmitter.js');
 
 import './src/mocks/react-native-async-storage.mock';
-import './src/mocks/react-native-reanimated.mock';
 import './src/mocks/react-native-themis.mock';
 import './src/mocks/react-native-device-info.mock';
