+++ libs/ui/package.json
@@ -56,14 +56,12 @@
     "react-native-permissions": "^3.6.1",
     "react-native-qrcode-scanner": "^1.5.5",
     "react-native-qrcode-svg": "^6.1.2",
-    "react-native-reanimated": "^2.13.0",
     "react-native-safe-area-context": "^4.2.5",
     "react-native-screens": "^3.18.0",
     "react-native-sha256": "^1.4.7",
     "react-native-svg": "^12.3.0",
     "react-native-themis": "^0.14.7",
     "react-native-toast-notifications": "^3.3.1",
-    "react-native-vision-camera": "2.15.1",
     "react-redux": "^8.0.1",
     "react-test-renderer": "^17.0.2",
     "redux-flipper": "^2.0.1",
@@ -74,7 +72,6 @@
     "ts-action-operators": "^9.1.2",
     "ts-prune": "^0.10.3",
     "typechain": "^8.1.1",
-    "vision-camera-code-scanner": "^0.2.0",
     "wasm-themis": "^0.14.5"
   },
   "resolutions": {
