--- libs/ui/src/mocks/react-native-reanimated.mock.js
@@ -1,11 +1,0 @@
-jest.mock('react-native-reanimated', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const Reanimated = require('react-native-reanimated/mock');
-
-  // The mock for `call` immediately calls the callback which is incorrect
-  // So we override it with a no-op
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  Reanimated.default.call = () => {};
-
-  return Reanimated;
-});
